+++ MoreFertilizers/MoreFertilizers/docs/Changelog.md
@@ -1,15 +1,16 @@
 ﻿Changelog
 =============
 
-
 #### Version 0.2.0 TODO items
 * Bush fertilizers
     - Handle adding to shops and other aquisition methods.
+    - Bountiful Bush fertilizer: actually implement.
 
 #### Version 0.2.0
 * Added new fertilizers:
     - Secret Joja Fertilizer: decreases growth time, has a chance of decreasing regrowth time, but forces the crop to be base quality.
     - Wisdom Fertilizer: increases farming XP by 10%.
+    - Rapid Bush Fertilizer: makes tea tree bushes grow 20% faster.
 
 * Adjusted previous fertilizers: Waterlogged now increases the growth speed of paddy crops a little.